add the regular mode checkers to proselint next-checkers
This seems to be working well. I would like to be able to use proselint
everywhere including for long comments in prog-modes, many of which are
already running multiple flycheck checkers that I would also like. This
function resets proselint's next-checkers in theory on every new buffer.
It allows proselint and the mode defined checkers to all be run.